toolkit/content/tests/widgets/test_moz_color_input.html

<!DOCTYPE HTML>
<html>

<head>
  <meta charset="utf-8">
  <title>moz-input-color Tests</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="module" src="chrome://global/content/elements/moz-input-color.mjs"></script>
  <link rel="stylesheet" href="chrome://mochikit/content/tests/SimpleTest/test.css" />
</head>

<body>
  <p id="display"></p>
  <div id="content">
    <moz-input-color color="#ffffff" prop-name="test-prop" data-l10n-id="color-input-test-label"></moz-input-color>
  </div>
  <pre id="test">
  <script class="testbody" type="application/javascript">
    add_task(async function testMozColorInput() {
      const colorInput = document.querySelector("moz-input-color");
      ok(colorInput, "color input element is rendered");

      let input = colorInput.shadowRoot.querySelector("input");
      is(input.value, "#ffffff", "color input has the correct initial value");
    });
    add_task(async function testMozColorInputEvents() {
      const colorInput = document.querySelector("moz-input-color");
      let input = colorInput.shadowRoot.querySelector("input");

      const pickedColor = new Promise((resolve) => {
        colorInput.addEventListener("color-picked", (event) => resolve(event.detail), { once: true });
      });
      input.value = "#0000ff";
      input.dispatchEvent(new Event("input"));
      let color = await pickedColor;
      is(color, "#0000ff", "color-picked event dispatches on input");
    });
  </script>
</pre>
</body>

</html>
